фикс globalSpeed и варнинг если ентитис больше чем 128

+++ entities.h
@@ -54,5 +54,5 @@ struct Projectile{
 std::vector<entity>& getEntities();
 
 void updateAll(std::vector<entity>& entities);
-void setGlobalSpeed(unsigned int speed);
+void setGlobalSpeed(double speed);
 #endif
